#413 reorganise zocalo tests
